# integration-api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Zinto Integration API
  version: 0.1.0
  summary: Bidirectional, tenant-safe API for Zinto partners
  description: |
    Integrate external systems with Zinto contacts, conversations, complete
    message history, channel delivery, notes, tags, and signed webhooks.
    Tenant identity always comes from the API key; clients never send company_id.
  x-api-base-url: https://crm.zinto.app/_integration-api/api/v1
servers:
  - url: https://crm.zinto.app/_integration-api
    description: Controlled preview on the Zinto VPS
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/webhooks:
    get:
      tags: [Webhooks]
      operationId: listWebhooks
      summary: List webhook endpoints without secrets
      x-required-scopes: [webhooks:manage]
      responses:
        "200":
          description: Webhook endpoint list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/WebhookEndpoint" }
                  meta: { $ref: "#/components/schemas/RequestMeta" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [Webhooks]
      operationId: createWebhook
      summary: Register an HTTPS endpoint; secret is returned once
      x-required-scopes: [webhooks:manage]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WebhookCreate" }
      responses:
        "201":
          description: Webhook endpoint and one-time signing secret
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    allOf:
                      - $ref: "#/components/schemas/WebhookEndpoint"
                      - type: object
                        properties:
                          secret: { type: string, pattern: '^whsec_[a-f0-9]{64}$' }
                  meta: { $ref: "#/components/schemas/RequestMeta" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "503": { $ref: "#/components/responses/ReadOnlyMode" }
  /api/v1/webhooks/{id}:
    parameters:
      - $ref: "#/components/parameters/ResourceId"
    delete:
      tags: [Webhooks]
      operationId: disableWebhook
      summary: Disable a webhook endpoint
      x-required-scopes: [webhooks:manage]
      responses:
        "204": { description: Endpoint disabled }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "503": { $ref: "#/components/responses/ReadOnlyMode" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: pcp_<64 lowercase hex characters>
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema: { type: string, pattern: '^\d+$' }
...
  responses:
...
    BadRequest:
      description: Validation or safety check failed
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Unauthorized:
      description: API key missing, invalid, inactive, or expired
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Forbidden:
      description: Required scope or source IP is not allowed
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    NotFound:
      description: Resource is absent or belongs to another company
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
    ReadOnlyMode:
      description: Write operations are temporarily disabled by READ_ONLY_MODE
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  schemas:
...
    RequestMeta:
      type: object
      required: [request_id]
      properties:
        request_id: { type: string, pattern: '^req_' }
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: array, items: {} }
            request_id: { type: string }
...
    WebhookCreate:
      type: object
      required: [url, event_types]
      additionalProperties: false
      properties:
        url: { type: string, format: uri, description: Public HTTPS endpoint }
        event_types:
          type: array
          minItems: 1
          maxItems: 20
          items: { $ref: "#/components/schemas/WebhookEventType" }
    WebhookEndpoint:
      type: object
      properties:
        id: { type: string }
        url: { type: string, format: uri }
        event_types: { type: array, items: { $ref: "#/components/schemas/WebhookEventType" } }
        active: { type: boolean }
        created_at: { type: string, format: date-time }
    WebhookEventType:
      type: string
      description: |
        Accepted subscription event emitted by the installed outbox migration.
        The corresponding source table must exist for optional ERP and Flow
        families to install their guarded triggers.
      enum:
        - contact.created
        - contact.updated
        - contact.deleted
        - conversation.created
        - conversation.updated
        - message.created
        - message.status.updated
        - note.created
        - note.updated
        - note.deleted
        - tag.attached
        - tag.detached
        - deal.created
        - deal.updated
        - deal.stage.changed
        - deal.deleted
        - pipeline.created
        - pipeline.updated
        - pipeline.deleted
        - pipeline.stage.created
        - pipeline.stage.updated
        - pipeline.stage.deleted
        - task.created
        - task.updated
        - task.completed
        - task.deleted
        - channel.connection.updated
        - erp.product.created
        - erp.product.updated
        - erp.product.deleted
        - erp.stock_level.created
        - erp.stock_level.updated
        - erp.stock_level.deleted
        - erp.stock_movement.created
        - erp.stock_movement.updated
        - erp.stock_movement.deleted
        - erp.stock_transfer.created
        - erp.stock_transfer.updated
        - erp.stock_transfer.deleted
        - erp.sales_order.created
        - erp.sales_order.updated
        - erp.sales_order.deleted
        - erp.supplier.created
        - erp.supplier.updated
        - erp.supplier.deleted
        - erp.purchase_order.created
        - erp.purchase_order.updated
        - erp.purchase_order.deleted
        - erp.invoice.created
        - erp.invoice.updated
        - erp.invoice.deleted
        - erp.invoice_payment.created
        - erp.invoice_payment.updated
        - erp.invoice_payment.deleted
        - flow.created
        - flow.updated
        - flow.deleted
        - flow.execution.started
        - flow.execution.updated
        - flow.execution.completed
        - flow.execution.failed
...